+++ mne_bids/physio/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["write_eyetrack_calibration"]
+
+from .eyetracking import write_eyetrack_calibration